requirements.txt
camelot-py==0.7.3
Django==3.0.4
django-appconf==1.0.3
django-compressor==2.3
django-extensions==2.2.3
pandas==1.0.1
pdfminer.six==20200124
PyPDF2==1.26.0
six==1.14.0
